refactor(cli): deduplicate concurrent issue processing and simplify helpers
- Extract shared goroutine/error-collection scaffold for assignee, comment, and transition commands into a single helper
- Reuse the existing flag-changed check in the string and bool flag accessors
- Reuse the existing CSV split helper when parsing search fields
- Drop the unused code-block tracking field from the Jira markdown renderer
- Check the buffer's last byte directly instead of stringifying it when rendering list items
- Short-circuit the boolean parser on the numeric form before lowercasing
